requirements.txt
altgraph==0.17
chromedriver-autoinstaller==0.2.2
future==0.18.2
macholib==1.14
numpy==1.20.1
pefile==2019.4.18
Pillow==8.1.0
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
PyYAML==5.4.1
selenium==3.141.0
six==1.15.0
urllib3==1.26.3
wxPython==4.1.1
